Create Rolling Year Total for SSRS Report Multi-Value Parameters
The procedure would be used in a SSRS report and calculates a rolling 12-month total. Dummy rows are added using a temp table method to allow for any missing months in the dataset returned also the multi-value parameters of the report are handled by using a string splitting funtions (dbo.splitstring) (see in SQLLibrary) and these are used in the WHERE clause which increased the performance of the stored procedure no end.