Create Page Life Expectancy Counter
The query gives returns the values of the Page Life Expectancy (PLE) on the server. This indicates how long a page stays in the buffer pool (in seconds). Higher values indicate a longer time in memory (i.e not going to disk) and better performance. "Traditionally" values over under 300 were considered problematic but with NUMA processor set ups this is no longer very a very meaningful value.